Relocate Exception classes
