+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -2,20 +2,17 @@ package nextstep.subway.line.application;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
-import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.SectionRequest;
+import nextstep.subway.line.exception.NoSuchLineException;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.dto.StationResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -36,10 +33,6 @@ public class LineService {
         return createLineResponse(persistLine);
     }
 
-    public List<Line> findLines() {
-        return lineRepository.findAll();
-    }
-
     public List<LineResponse> findLineResponses() {
         final List<Line> persistLines = lineRepository.findAll();
         return persistLines.stream()
@@ -48,7 +41,7 @@ public class LineService {
     }
 
     public Line findLineById(Long id) {
-        return lineRepository.findById(id).orElseThrow(RuntimeException::new);
+        return lineRepository.findById(id).orElseThrow(()-> new NoSuchLineException("존재하지 않는 노선입니다"));
     }
 
     public LineResponse findLineResponseById(Long id) {
@@ -57,7 +50,7 @@ public class LineService {
     }
 
     public void updateLine(Long id, LineRequest lineUpdateRequest) {
-        final Line persistLine = lineRepository.findById(id).orElseThrow(RuntimeException::new);
+        final Line persistLine = lineRepository.findById(id).orElseThrow(()-> new NoSuchLineException("존재하지 않는 노선입니다"));
         persistLine.update(new Line(lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
     }
