+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,6 +1,7 @@
 package nextstep.subway.line.domain;
 
 import nextstep.subway.common.BaseEntity;
+import nextstep.subway.line.exception.InvalidStationException;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
