+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.domain;
 
+import nextstep.subway.line.exception.InvalidSectionOperationException;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.CascadeType;
