+++ src/main/java/nextstep/subway/line/exception/InvalidSectionDistanceException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.exception;
 
 public class InvalidSectionDistanceException extends RuntimeException{
     public InvalidSectionDistanceException(String msg){
