+++ src/main/java/nextstep/subway/line/exception/InvalidSectionOperationException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.exception;
 
 public class InvalidSectionOperationException extends RuntimeException{
     public InvalidSectionOperationException(String msg){
