+++ src/main/java/nextstep/subway/line/exception/InvalidStationException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.exception;
 
 public class InvalidStationException extends RuntimeException{
     public InvalidStationException(String msg){
