+++ src/main/java/nextstep/subway/line/exception/NoSuchLineException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.line.exception;
+
+public class NoSuchLineException extends RuntimeException{
+    public NoSuchLineException(String msg){
+        super(msg);
+    }
+}
