+++ src/main/java/nextstep/subway/station/application/NoSuchStationException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.station.application;
+
+public class NoSuchStationException extends RuntimeException{
+    public NoSuchStationException(String msg){
+        super(msg);
+    }
+}
