+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,5 +1,7 @@
 package nextstep.subway.line.domain;
 
+import nextstep.subway.line.exception.InvalidSectionOperationException;
+import nextstep.subway.line.exception.InvalidStationException;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
